Better interface for Atomic ranging
AtomicRanger is now separate from AtomicComponent. It is still set
during instantiation in ComponentCollection, but there is now a field in
ComponentCollection that allows one to choose the RangingStrategy used.
