+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/Variable.scala
@@ -15,10 +15,13 @@ package com.cra.figaro.algorithm.factored.factors
 
 import com.cra.figaro.algorithm._
 import com.cra.figaro.language._
+
 import scala.collection.mutable.Map
-import com.cra.figaro.algorithm.lazyfactored.{ LazyValues, Extended, ValueSet }
+import com.cra.figaro.algorithm.lazyfactored.{Extended, LazyValues, ValueSet}
 import com.cra.figaro.algorithm.structured._
+import com.cra.figaro.algorithm.structured.strategy.range.ValuesRanger
 import com.cra.figaro.library.collection.MakeArray
+
 import scala.collection.mutable.HashMap
 
 /**
@@ -99,9 +102,9 @@ object Variable {
     case chain: Chain[_, T] => new ChainComponent(problem, chain)
     case makeArray: MakeArray[_] => new MakeArrayComponent(problem, makeArray).asInstanceOf[ProblemComponent[T]]
     case apply: Apply[T] => new ApplyComponent(problem, apply)
-    // This is subtle: we must use the atomic component type that generates a distribution based on the range given,
+    // This is subtle: we must use the atomic ranger type that generates a distribution based on the range given,
     // rather than one that generates its own range
-    case atomic: Atomic[T] => new ValuesAtomicComponent(problem, atomic)
+    case atomic: Atomic[T] => new AtomicComponent(problem, atomic, new ValuesRanger(atomic, cc))
     case _ => new ProblemComponent(problem, elem)
   }
