+++ Figaro/src/main/scala/com/cra/figaro/algorithm/structured/strategy/refine/DecompositionStrategy.scala
@@ -204,7 +204,7 @@ abstract class DecompositionStrategy(collection: ComponentCollection, done: muta
     // An atomic element has no args; simply generate its range
     generateRange(atomicComp)
     // Decide if the component is fully enumerated/refined based on the ranging strategy used
-    atomicComp.fullyRefined = atomicComp.fullyRefinable()
+    atomicComp.fullyRefined = atomicComp.ranger.fullyRefinable()
     atomicComp.fullyEnumerated = atomicComp.fullyRefined
   }
